Mirror the entity validator split for policies
