+++ test/elixir/hologram/policy/validator_test.exs
@@ -1,143 +1,13 @@
-defmodule Hologram.Policy.CompilerTest do
+defmodule Hologram.Policy.ValidatorTest do
   use Hologram.Test.BasicCase, async: true
 
-  import Hologram.Policy.Compiler
+  import Hologram.Policy.Validator
 
-  alias Hologram.Auth.RoleGrant
   alias Hologram.Test.Fixtures.Entity.Module1
   alias Hologram.Test.Fixtures.Entity.Module13
   alias Hologram.Test.Fixtures.Entity.Module14
   alias Hologram.Test.Fixtures.Entity.Module2
-  alias Hologram.Test.Fixtures.Entity.Module4
-  alias Hologram.Test.Fixtures.Policy
   alias Hologram.Test.Fixtures.Role
-
-  describe "build/1" do
-    test "returns empty map for entity type without policy declarations" do
-      assert build(Module1) == %{}
-    end
-
-    test "builds rules per operation, keeping declaration order" do
-      assert build(Policy.Module1) == %{
-               archive: [
-                 %{predicates: [{:author_id, :==, {:actor}}], to: nil, via: nil}
-               ],
-               delete: [
-                 %{predicates: [], to: [{:rel, :parent, [:admin]}], via: nil}
-               ],
-               manage_roles: [
-                 %{predicates: [], to: [{:own, [:owner]}], via: nil}
-               ],
-               publish: [
-                 %{predicates: [], to: nil, via: :parent}
-               ],
-               read: [
-                 %{predicates: [{:public, :==, true}], to: nil, via: nil},
-                 %{
-                   predicates: [],
-                   to: [{:own, [:viewer]}, {:type, Policy.Module2, [:admin]}],
-                   via: nil
-                 }
-               ],
-               update: [
-                 %{predicates: [{:priority, :>=, 3}], to: [{:own, [:editor, :owner]}], via: nil}
-               ]
-             }
-    end
-  end
-
-  describe "build/1 with global role references" do
-    test "expands a role module reference to every role carrying it" do
-      defmodule GlobalReferenceFixture do
-        use Hologram.Entity
-
-        allow :archive, to: Hologram.Test.Fixtures.Role.Module1
-      end
-
-      assert build(GlobalReferenceFixture)[:archive] == [
-               %{
-                 predicates: [],
-                 to: [{:global, [Role.Module1, Role.Module2]}],
-                 via: nil
-               }
-             ]
-    end
-
-    test "keeps own and global references of one line apart" do
-      defmodule MixedReferenceFixture do
-        use Hologram.Entity
-
-        role :viewer
-
-        allow :read, to: [:viewer, Hologram.Test.Fixtures.Role.Module2]
-      end
-
-      assert build(MixedReferenceFixture)[:read] == [
-               %{
-                 predicates: [],
-                 to: [{:own, [:viewer]}, {:global, [Role.Module2]}],
-                 via: nil
-               }
-             ]
-    end
-  end
-
-  describe "build/1 for the grant store" do
-    test "grants sight of own grants, and of others' grants to read-grants role holders" do
-      assert build(RoleGrant) == %{
-               read: [
-                 %{predicates: [{:user_id, :==, {:actor}}], to: nil, via: nil},
-                 %{
-                   predicates: [{:resource_type, :==, :test_fixtures_policy_module1}],
-                   to: [{:resource, Policy.Module1, [:owner]}],
-                   via: nil
-                 },
-                 %{
-                   predicates: [{:resource_type, :==, :test_fixtures_policy_module2}],
-                   to: [{:resource, Policy.Module2, [:member]}],
-                   via: nil
-                 }
-               ]
-             }
-    end
-  end
-
-  describe "dead_entity_types/1" do
-    test "returns the entity types declaring no allow lines, sorted" do
-      assert dead_entity_types([Policy.Module1, Module14, Module2, Module1, Module4]) == [
-               Module1,
-               Module4
-             ]
-    end
-
-    test "returns empty list when every entity type declares an allow line" do
-      assert dead_entity_types([Policy.Module1, Policy.Module2]) == []
-    end
-
-    test "never returns the grant store" do
-      assert dead_entity_types([RoleGrant]) == []
-    end
-  end
-
-  describe "manage_roles_qualifying_roles/1" do
-    test "returns the expanded own roles of the manage_roles rules" do
-      assert manage_roles_qualifying_roles(Policy.Module1) == [:owner]
-    end
-
-    test "returns empty list when the entity type declares no manage_roles rule" do
-      assert manage_roles_qualifying_roles(Policy.Module2) == []
-    end
-  end
-
-  describe "read_grants_roles/1" do
-    test "returns the expanded own roles of the read_grants rules" do
-      assert read_grants_roles(Policy.Module2) == [:member]
-    end
-
-    test "defaults to the roles qualifying to manage grants" do
-      assert read_grants_roles(Policy.Module1) == [:owner]
-    end
-  end
 
   describe "validate_model!/1" do
     test "returns :ok for empty model" do
@@ -282,7 +152,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "multiple user entity designations in the data model: Hologram.Policy.CompilerTest.InlinePolicyFixture27, Hologram.Policy.CompilerTest.InlinePolicyFixture28 - exactly one entity type can be designated with use Hologram.Entity, user: true"
+        "multiple user entity designations in the data model: Hologram.Policy.ValidatorTest.InlinePolicyFixture27, Hologram.Policy.ValidatorTest.InlinePolicyFixture28 - exactly one entity type can be designated with use Hologram.Entity, user: true"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture28, InlinePolicyFixture27])
@@ -297,7 +167,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "Hologram.Policy.CompilerTest.InlinePolicyFixture29 declares roles or policy grant references, but no entity type is designated as the user entity - add use Hologram.Entity, user: true to your user module"
+        "Hologram.Policy.ValidatorTest.InlinePolicyFixture29 declares roles or policy grant references, but no entity type is designated as the user entity - add use Hologram.Entity, user: true to your user module"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture29])
@@ -312,7 +182,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "Hologram.Policy.CompilerTest.InlinePolicyFixture30 declares roles or policy grant references, but no entity type is designated as the user entity - add use Hologram.Entity, user: true to your user module"
+        "Hologram.Policy.ValidatorTest.InlinePolicyFixture30 declares roles or policy grant references, but no entity type is designated as the user entity - add use Hologram.Entity, user: true to your user module"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture30])
@@ -329,7 +199,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid predicate for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture8 - user_id() requires a uuid attribute - attribute :title in Hologram.Policy.CompilerTest.InlinePolicyFixture8 has type :string"
+        "invalid predicate for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture8 - user_id() requires a uuid attribute - attribute :title in Hologram.Policy.ValidatorTest.InlinePolicyFixture8 has type :string"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture8])
@@ -347,7 +217,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "unknown role :publisher in the to option of allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture11 - declared roles are: :editor, :owner"
+        "unknown role :publisher in the to option of allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture11 - declared roles are: :editor, :owner"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture11])
@@ -362,7 +232,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid to option {Hologram.Reflection, :editor} for allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture15 - Hologram.Reflection is not an entity type module"
+        "invalid to option {Hologram.Reflection, :editor} for allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture15 - Hologram.Reflection is not an entity type module"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture15])
@@ -377,7 +247,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "unknown role :publisher in the to option of allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture16 - declared roles of Hologram.Test.Fixtures.Entity.Module13 are: :editor, :owner"
+        "unknown role :publisher in the to option of allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture16 - declared roles of Hologram.Test.Fixtures.Entity.Module13 are: :editor, :owner"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture16])
@@ -394,7 +264,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "unknown relationship :project in the to option of allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture18 - declared relationships are: :parent"
+        "unknown relationship :project in the to option of allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture18 - declared relationships are: :parent"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture18])
@@ -411,7 +281,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid to option {:children, :editor} for allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture19 - relationship :children is to-many, but a role reference requires a to-one relationship"
+        "invalid to option {:children, :editor} for allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture19 - relationship :children is to-many, but a role reference requires a to-one relationship"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture19])
@@ -428,7 +298,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "unknown role :publisher in the to option of allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture20 - declared roles of Hologram.Test.Fixtures.Entity.Module13 are: :editor, :owner"
+        "unknown role :publisher in the to option of allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture20 - declared roles of Hologram.Test.Fixtures.Entity.Module13 are: :editor, :owner"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture20])
@@ -445,7 +315,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid to option \"owner\" for allow :update in Hologram.Policy.CompilerTest.InlinePolicyFixture12 - the to option must be a role name, a {module, role} or {relationship, role} tuple, or a non-empty list of them"
+        "invalid to option \"owner\" for allow :update in Hologram.Policy.ValidatorTest.InlinePolicyFixture12 - the to option must be a role name, a {module, role} or {relationship, role} tuple, or a non-empty list of them"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InlinePolicyFixture12])
@@ -462,7 +332,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "unknown relationship :project in the via option of allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture22 - declared relationships are: :parent"
+        "unknown relationship :project in the via option of allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture22 - declared relationships are: :parent"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture22])
@@ -479,7 +349,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid via option :children for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture23 - relationship :children is to-many, but delegation requires a to-one relationship"
+        "invalid via option :children for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture23 - relationship :children is to-many, but delegation requires a to-one relationship"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture23])
@@ -496,7 +366,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid via option \"parent\" for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture24 - the via option must be a relationship name"
+        "invalid via option \"parent\" for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture24 - the via option must be a relationship name"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture24])
@@ -507,7 +377,7 @@ defmodule Hologram.Policy.CompilerTest do
       defmodule InlinePolicyFixture25 do
         use Hologram.Entity
 
-        relationship :peer, Hologram.Policy.CompilerTest.InlinePolicyFixture26, optional: true
+        relationship :peer, Hologram.Policy.ValidatorTest.InlinePolicyFixture26, optional: true
 
         allow :read, via: :peer
       end
@@ -515,7 +385,7 @@ defmodule Hologram.Policy.CompilerTest do
       defmodule InlinePolicyFixture26 do
         use Hologram.Entity
 
-        relationship :peer, Hologram.Policy.CompilerTest.InlinePolicyFixture25, optional: true
+        relationship :peer, Hologram.Policy.ValidatorTest.InlinePolicyFixture25, optional: true
 
         allow :read, via: :peer
       end
@@ -523,7 +393,7 @@ defmodule Hologram.Policy.CompilerTest do
       expected_msg =
         normalize_newlines("""
         cyclic policy delegation for allow :read - a via chain can't return to the entity type it starts from:
-          * Hologram.Policy.CompilerTest.InlinePolicyFixture25 (via :peer) -> Hologram.Policy.CompilerTest.InlinePolicyFixture26 (via :peer) -> Hologram.Policy.CompilerTest.InlinePolicyFixture25\
+          * Hologram.Policy.ValidatorTest.InlinePolicyFixture25 (via :peer) -> Hologram.Policy.ValidatorTest.InlinePolicyFixture26 (via :peer) -> Hologram.Policy.ValidatorTest.InlinePolicyFixture25\
         """)
 
       assert_error Hologram.CompileError, expected_msg, fn ->
@@ -541,7 +411,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid predicate for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture3 - unknown attribute :published in Hologram.Policy.CompilerTest.InlinePolicyFixture3 - known attributes: :created_at, :id, :title, :updated_at"
+        "invalid predicate for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture3 - unknown attribute :published in Hologram.Policy.ValidatorTest.InlinePolicyFixture3 - known attributes: :created_at, :id, :title, :updated_at"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture3])
@@ -558,7 +428,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid predicate for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture4 - :parent is a relationship in Hologram.Policy.CompilerTest.InlinePolicyFixture4 - only attributes can be filtered - filter its reference via :parent_id"
+        "invalid predicate for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture4 - :parent is a relationship in Hologram.Policy.ValidatorTest.InlinePolicyFixture4 - only attributes can be filtered - filter its reference via :parent_id"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture4])
@@ -575,7 +445,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid predicate for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture5 - unknown operator :like in the filter predicate for attribute :title - supported operators: :!=, :<, :<=, :==, :>, :>=, :in, :not_in"
+        "invalid predicate for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture5 - unknown operator :like in the filter predicate for attribute :title - supported operators: :!=, :<, :<=, :==, :>, :>=, :in, :not_in"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture5])
@@ -592,7 +462,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid predicate for allow :read in Hologram.Policy.CompilerTest.InlinePolicyFixture6 - operator :>= requires a numeric or temporal attribute - attribute :parent_id in Hologram.Policy.CompilerTest.InlinePolicyFixture6 has type :uuid"
+        "invalid predicate for allow :read in Hologram.Policy.ValidatorTest.InlinePolicyFixture6 - operator :>= requires a numeric or temporal attribute - attribute :parent_id in Hologram.Policy.ValidatorTest.InlinePolicyFixture6 has type :uuid"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([InlinePolicyFixture6])
@@ -619,7 +489,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid to option Hologram.Test.Fixtures.Entity.Module1 for allow :read in Hologram.Policy.CompilerTest.InvalidGlobalReferenceFixture - Hologram.Test.Fixtures.Entity.Module1 is not a role module (define it with use Hologram.Role)"
+        "invalid to option Hologram.Test.Fixtures.Entity.Module1 for allow :read in Hologram.Policy.ValidatorTest.InvalidGlobalReferenceFixture - Hologram.Test.Fixtures.Entity.Module1 is not a role module (define it with use Hologram.Role)"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_model!([Module14, InvalidGlobalReferenceFixture])
@@ -650,7 +520,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "invalid extends target Hologram.Test.Fixtures.Entity.Module1 in use Hologram.Role for Hologram.Policy.CompilerTest.InvalidTargetRoleFixture - extends targets must be modules defined with use Hologram.Role"
+        "invalid extends target Hologram.Test.Fixtures.Entity.Module1 in use Hologram.Role for Hologram.Policy.ValidatorTest.InvalidTargetRoleFixture - extends targets must be modules defined with use Hologram.Role"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_role_modules!([InvalidTargetRoleFixture])
@@ -659,17 +529,17 @@ defmodule Hologram.Policy.CompilerTest do
 
     test "rejects an extension cycle" do
       defmodule CyclicRoleFixture1 do
-        use Hologram.Role, extends: Hologram.Policy.CompilerTest.CyclicRoleFixture2
+        use Hologram.Role, extends: Hologram.Policy.ValidatorTest.CyclicRoleFixture2
       end
 
       defmodule CyclicRoleFixture2 do
-        use Hologram.Role, extends: Hologram.Policy.CompilerTest.CyclicRoleFixture1
+        use Hologram.Role, extends: Hologram.Policy.ValidatorTest.CyclicRoleFixture1
       end
 
       expected_msg =
         normalize_newlines("""
         cyclic role extension - an extends chain can't return to the role it starts from:
-          * Hologram.Policy.CompilerTest.CyclicRoleFixture1 -> Hologram.Policy.CompilerTest.CyclicRoleFixture2 -> Hologram.Policy.CompilerTest.CyclicRoleFixture1\
+          * Hologram.Policy.ValidatorTest.CyclicRoleFixture1 -> Hologram.Policy.ValidatorTest.CyclicRoleFixture2 -> Hologram.Policy.ValidatorTest.CyclicRoleFixture1\
         """)
 
       assert_error Hologram.CompileError, expected_msg, fn ->
@@ -683,7 +553,7 @@ defmodule Hologram.Policy.CompilerTest do
       end
 
       expected_msg =
-        "role module name Hologram.Policy.CompilerTest.ThisIsAVeryLongRoleModuleNameUsedToExceedThePostgresEnumLabelLimit is too long to store as a grant value (95 bytes, limit 63) - shorten the module name"
+        "role module name Hologram.Policy.ValidatorTest.ThisIsAVeryLongRoleModuleNameUsedToExceedThePostgresEnumLabelLimit is too long to store as a grant value (96 bytes, limit 63) - shorten the module name"
 
       assert_error Hologram.CompileError, expected_msg, fn ->
         validate_role_modules!([
