+++ test/elixir/hologram/policy_test.exs
@@ -1,7 +1,20 @@
 defmodule Hologram.PolicyTest do
   use Hologram.Test.BasicCase, async: true
 
-  alias Hologram.Policy.Compiler
+  import Hologram.Policy,
+    only: [
+      build: 1,
+      dead_entity_types: 1,
+      manage_roles_qualifying_roles: 1,
+      read_grants_roles: 1
+    ]
+
+  alias Hologram.Auth.RoleGrant
+  alias Hologram.Test.Fixtures.Entity.Module1
+  alias Hologram.Test.Fixtures.Entity.Module14
+  alias Hologram.Test.Fixtures.Entity.Module2
+  alias Hologram.Test.Fixtures.Entity.Module4
+  alias Hologram.Test.Fixtures.Policy
   alias Hologram.Test.Fixtures.Role
 
   describe "allow/2" do
@@ -41,7 +54,7 @@ defmodule Hologram.PolicyTest do
         use ActorPolicyFixture
       end
 
-      assert Compiler.build(ActorEntityFixture)[:archive] == [
+      assert build(ActorEntityFixture)[:archive] == [
                %{predicates: [{:id, :==, {:actor}}], to: nil, via: nil}
              ]
     end
@@ -60,7 +73,7 @@ defmodule Hologram.PolicyTest do
         use GlobalReferencePolicyFixture
       end
 
-      assert Compiler.build(GlobalReferenceEntityFixture)[:read] == [
+      assert build(GlobalReferenceEntityFixture)[:read] == [
                %{predicates: [], to: [{:global, [Role.Module1, Role.Module2]}], via: nil}
              ]
     end
@@ -203,6 +216,133 @@ defmodule Hologram.PolicyTest do
           use TargetPolicyFixture
         end
       end
+    end
+  end
+
+  describe "build/1" do
+    test "returns empty map for entity type without policy declarations" do
+      assert build(Module1) == %{}
+    end
+
+    test "builds rules per operation, keeping declaration order" do
+      assert build(Policy.Module1) == %{
+               archive: [
+                 %{predicates: [{:author_id, :==, {:actor}}], to: nil, via: nil}
+               ],
+               delete: [
+                 %{predicates: [], to: [{:rel, :parent, [:admin]}], via: nil}
+               ],
+               manage_roles: [
+                 %{predicates: [], to: [{:own, [:owner]}], via: nil}
+               ],
+               publish: [
+                 %{predicates: [], to: nil, via: :parent}
+               ],
+               read: [
+                 %{predicates: [{:public, :==, true}], to: nil, via: nil},
+                 %{
+                   predicates: [],
+                   to: [{:own, [:viewer]}, {:type, Policy.Module2, [:admin]}],
+                   via: nil
+                 }
+               ],
+               update: [
+                 %{predicates: [{:priority, :>=, 3}], to: [{:own, [:editor, :owner]}], via: nil}
+               ]
+             }
+    end
+  end
+
+  describe "build/1 with global role references" do
+    test "expands a role module reference to every role carrying it" do
+      defmodule GlobalReferenceFixture do
+        use Hologram.Entity
+
+        allow :archive, to: Hologram.Test.Fixtures.Role.Module1
+      end
+
+      assert build(GlobalReferenceFixture)[:archive] == [
+               %{
+                 predicates: [],
+                 to: [{:global, [Role.Module1, Role.Module2]}],
+                 via: nil
+               }
+             ]
+    end
+
+    test "keeps own and global references of one line apart" do
+      defmodule MixedReferenceFixture do
+        use Hologram.Entity
+
+        role :viewer
+
+        allow :read, to: [:viewer, Hologram.Test.Fixtures.Role.Module2]
+      end
+
+      assert build(MixedReferenceFixture)[:read] == [
+               %{
+                 predicates: [],
+                 to: [{:own, [:viewer]}, {:global, [Role.Module2]}],
+                 via: nil
+               }
+             ]
+    end
+  end
+
+  describe "build/1 for the grant store" do
+    test "grants sight of own grants, and of others' grants to read-grants role holders" do
+      assert build(RoleGrant) == %{
+               read: [
+                 %{predicates: [{:user_id, :==, {:actor}}], to: nil, via: nil},
+                 %{
+                   predicates: [{:resource_type, :==, :test_fixtures_policy_module1}],
+                   to: [{:resource, Policy.Module1, [:owner]}],
+                   via: nil
+                 },
+                 %{
+                   predicates: [{:resource_type, :==, :test_fixtures_policy_module2}],
+                   to: [{:resource, Policy.Module2, [:member]}],
+                   via: nil
+                 }
+               ]
+             }
+    end
+  end
+
+  describe "dead_entity_types/1" do
+    test "returns the entity types declaring no allow lines, sorted" do
+      assert dead_entity_types([Policy.Module1, Module14, Module2, Module1, Module4]) == [
+               Module1,
+               Module4
+             ]
+    end
+
+    test "returns empty list when every entity type declares an allow line" do
+      assert dead_entity_types([Policy.Module1, Policy.Module2]) == []
+    end
+
+    test "never returns the grant store" do
+      assert dead_entity_types([RoleGrant]) == []
+    end
+  end
+
+  describe "manage_roles_qualifying_roles/1" do
+    test "returns the expanded own roles of the manage_roles rules" do
+      assert manage_roles_qualifying_roles(Policy.Module1) == [:owner]
+    end
+
+    test "returns empty list when the entity type declares no manage_roles rule" do
+      assert manage_roles_qualifying_roles(Policy.Module2) == []
+    end
+  end
+
+  describe "read_grants_roles/1" do
+    test "returns the expanded own roles of the read_grants rules" do
+      assert read_grants_roles(Policy.Module2) == [:member]
+    end
+
+    test "defaults to the roles qualifying to manage grants" do
+      assert read_grants_roles(Policy.Module1) == [:owner]
     end
   end
 end
